fix error: cast from 'void*' to 'long int' loses precision [-fpermissive]

+++ Lab07/lab07.cpp
@@ -157,26 +157,26 @@ void two(long number)              // 345678
 
    // change pointerFunction in main() to point to pass
    pLong = (long *) &bow;
-   while(*pLong != (long)fail) //This time we look for the function, so convert the addres to long pointer again
+   while(*pLong != (long long)fail) //This time we look for the function, so convert the addres to long pointer again
    {
       pLong++; //increment the pointer address by one byte
    }
    cout << *pLong << endl;
    //once found, just change the value in that address
-   *pLong = (long)pass;
+   *pLong = (long long)pass;
 
    cout << *pLong << endl;
    // change message in main() to point to passMessage
    
    pLong = (long *) &bow;
-   while(*pLong != (long)failMessage) //failMessage is a global variable so it shiuold be available here.
+   while(*pLong != (long long)failMessage) //failMessage is a global variable so it shiuold be available here.
    {
       pLong++; //increment the pointer address by one byte
    }
    pChar = (char *) *pLong; //Convert the pLong value to char and be able to print and check
    cout << string(pChar)<< endl;
    //once found, just change the value in that address
-   *pLong = (long)passMessage; //convert the pointer to long and store it in pLong
+   *pLong = (long long)passMessage; //convert the pointer to long and store it in pLong
                               //Since pLong is pointing to the same value than message in main is pointing
                               //It should change the message value too.
